dinero en centimos, bug historial y comprobaciones para apostar (#20)
* elimino archivos de db

* parche apostar a eventos acabados/cancelados

* solo se permiten variables que existan en las formulas

* cargar mas y lupa de index arregladas

* cambio de la representacion del dinero a centimos

* arreglo bug historico

---------

+++ src/main/java/es/ucm/fdi/iw/model/User.java
@@ -55,8 +55,8 @@ public class User implements Transferable<User.Transfer> {
     private boolean enabled;
     private String roles; // split by ',' to separate roles
 
-    private double dineroDisponible;
-    private double dineroRetenido;
+    private int dineroDisponible; // En centimos
+    private int dineroRetenido; // En centimos
 
     @ManyToMany
     @JoinTable(
